Address PR
- Merge in master
- `/radio` => `/radios`. For some reason `/radio/${id}` just gave me a status=0? If the original one works that'd be great
- Move radio search to empty page/radio list
- Edit button for existing entries instead

+++ server/subsonic/playlists.go
@@ -165,9 +165,9 @@ func (api *Router) buildPlaylist(p model.Playlist) *responses.Playlist {
 	pls.Id = p.ID
 	pls.Name = p.Name
 	pls.Comment = p.Comment
-	pls.SongCount = p.SongCount
+	pls.SongCount = int32(p.SongCount)
 	pls.Owner = p.OwnerName
-	pls.Duration = int(p.Duration)
+	pls.Duration = int32(p.Duration)
 	pls.Public = p.Public
 	pls.Created = p.CreatedAt
 	pls.Changed = p.UpdatedAt
